Move device geometry headers into SFM
