+++ gtsam/sfm/cuda/internal/DeviceGeometryTypes.h
