+++ gtsam/sfm/tests/testCudaSfm.cpp
@@ -22,11 +22,11 @@
 #include <gtsam/linear/cuda/internal/BlockOrdering.h>
 #include <gtsam/nonlinear/BatchFactor.h>
 #include <gtsam/nonlinear/LevenbergMarquardtOptimizer.h>
-#include <gtsam/nonlinear/cuda/internal/DeviceGeometryKernels.h>
-#include <gtsam/nonlinear/cuda/internal/DeviceGeometryTypes.h>
 #include <gtsam/sfm/SfmData.h>
 #include <gtsam/slam/GeneralSFMFactor.h>
 #include <gtsam/sfm/cuda/SfmLevenbergMarquardt.h>
+#include <gtsam/sfm/cuda/internal/DeviceGeometryKernels.h>
+#include <gtsam/sfm/cuda/internal/DeviceGeometryTypes.h>
 #include <gtsam/sfm/cuda/internal/SfmDenseSchurSolver.h>
 #include <gtsam/sfm/cuda/internal/SfmProjectionBatch.h>
 #include <gtsam/sfm/cuda/internal/SfmProjectionLinearization.h>
